Imagine you are having dinner in a very fancy restaurant, but unfortunately you don't have a lot of money with you. You want to have a main course, a dessert and a drink, but all that together shouldn't cost more than $30.
The names of the main courses, desserts and drinks are stored in the lists main_courses, desserts and drinks respectively. The corresponding prices can be found in the lists price_main_courses, price_desserts and price_drinks.

Consider each possible combination of a main course, dessert and a drink from those offered by the restaurant and print out only those meals that satisfy your budget, along with their total costs.